# openapi.yaml
...
openapi: 3.1.0
info:
  title: Lentera — LMS Wakaf Salman API
  version: 1.0.0
  description: |-
    Kontrak berdasarkan implementasi Flask di repository. Gunakan Bearer JWT dari POST /api/auth/login. Logout tercatat melalui POST /api/auth/logout; token baru yang terkait sesi login tersebut ditolak setelah logout. Role dan status akun dibaca ulang dari database pada setiap request.

    Path learning memakai /api/course (singular); assessment memakai /api/courses (plural). ID berasal dari database, bukan ID contoh/frontend. Daftar katalog/user/enrollment tidak memiliki pagination; hanya activity mendukung pagination.

    Payload JSON memakai Content-Type: application/json. Field request yang tidak dikenal ditolak kecuali login/register yang mengabaikan field tambahan. Input teks umumnya di-trim. Global request limit 3 MiB; endpoint profil memakai batas 2 MiB + 64 KiB. Error aplikasi umumnya {message}; error framework/proxy (termasuk 413 di luar auth) dapat berupa HTML. Timestamp identitas/learning dapat tanpa timezone, sedangkan timestamp assessment/discussion memakai UTC dengan Z.

    Endpoint Development bersifat opsional dan selalu tidak tersedia pada production. Spesifikasi ini tidak menambahkan route HTTP baru atau Swagger UI.
servers:
- url: http://localhost:5000
  description: Flask development lokal
- url: http://localhost:8000
  description: Docker API lokal; sesuaikan API_PORT
- url: https://{hostname}
...
  variables:
    hostname:
      default: api.example.com
...
security:
- bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Registrasi akun user
      operationId: post_api_auth_register
      responses:
        '201':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Registrasi berhasil
        '400':
          $ref: '#/components/responses/Error400'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Login dan dapatkan JWT
      operationId: post_api_auth_login
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '413':
          $ref: '#/components/responses/Error413'
        '503':
          $ref: '#/components/responses/Error503'
      security: []
      description: Token berlaku 8 jam. Login mencatat activity; kegagalan commit tidak menghasilkan token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /api/auth/me:
    get:
      tags:
      - Auth
      summary: Profil akun sendiri
      operationId: get_api_auth_me
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      description: Mengembalikan object user langsung, tanpa pembungkus user.
    patch:
      tags:
      - Auth
      summary: Ubah profil atau upload foto
      operationId: patch_api_auth_me
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                required:
                - message
                - user
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '503':
          $ref: '#/components/responses/Error503'
      description: Hanya profil sendiri. JSON atau multipart/form-data; kirim satu file dengan nama foto. Request
        multipart maksimal 2 MiB + 64 KiB. File maksimal 2 MiB. Upload foto tanpa field teks juga diperbolehkan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProfileUpload'
  /api/auth/profile-photos/{filename}:
    get:
      tags:
      - Auth
      summary: Ambil foto profil PNG
      operationId: get_api_auth_profile_photos_filename
      responses:
        '200':
          description: Foto profil publik; cache maksimal 86400 detik.
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Nama file tidak valid atau file tidak ditemukan.
          content:
            text/html:
              schema:
                type: string
      security: []
      parameters:
      - name: filename
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-f0-9]{32}\.png$
  /api/auth/admin-test:
    get:
      tags:
      - Auth
      summary: Verifikasi akses admin
      operationId: get_api_auth_admin_test
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      id_user:
                        type: integer
                        format: int64
                        minimum: 1
                      nama:
                        type: string
                      role:
                        type: string
                        enum:
                        - admin
                        - user
                    required:
                    - id_user
                    - nama
                    - role
                required:
                - message
                - user
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-required-role: admin
      description: 'Hanya admin aktif. '
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT HS256 dari login; Authorization: Bearer <token>. Berlaku 8 jam.'
  responses:
    Error400:
      description: Payload atau parameter tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error401:
      description: Token hilang/invalid/kedaluwarsa, akun tidak ditemukan, atau kredensial login salah.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: Akun nonaktif, role tidak sesuai, bukan pemilik, atau prasyarat akses belum terpenuhi.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Error409:
      description: Konflik data, relasi, konfigurasi test, atau status operasi.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error413:
      description: Request terlalu besar.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error503:
      description: Penyimpanan sementara tidak tersedia.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
          example: Operasi berhasil
      required:
      - message
    Error:
      type: object
      properties:
        message:
          type: string
          example: Data tidak valid
      required:
      - message
    User:
      type: object
      properties:
        id_user:
          type: integer
          format: int64
          minimum: 1
        nama:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 150
        user_profile:
          type: object
          properties:
            no_hp:
              anyOf:
              - type: string
                maxLength: 30
              - type: 'null'
          required:
          - no_hp
        divisi:
          anyOf:
          - type: string
          - type: 'null'
        jabatan:
          anyOf:
          - type: string
          - type: 'null'
        role:
          type: string
          enum:
          - admin
          - user
        status:
          type: string
          enum:
          - aktif
          - nonaktif
        foto_profil:
          anyOf:
          - type: string
            description: URL foto lengkap atau string kosong jika dihapus.
          - type: 'null'
        tanggal_dibuat:
          anyOf:
          - type: string
            description: Tanggal/waktu ISO 8601 tanpa timezone, mengikuti waktu server.
            example: '2026-10-06T10:30:00'
          - type: 'null'
        terakhir_login:
          anyOf:
          - type: string
            description: Tanggal/waktu ISO 8601 tanpa timezone, mengikuti waktu server.
            example: '2026-10-06T10:30:00'
          - type: 'null'
      required:
      - id_user
      - nama
      - email
      - user_profile
      - divisi
      - jabatan
      - role
      - status
      - foto_profil
      - tanggal_dibuat
      - terakhir_login
    LoginUser:
      type: object
      properties:
        id_user:
          type: integer
          format: int64
          minimum: 1
        nama:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - admin
          - user
      required:
      - id_user
      - nama
      - email
      - role
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
          description: JWT HS256; kedaluwarsa 8 jam setelah login.
        user:
          $ref: '#/components/schemas/LoginUser'
      required:
      - message
      - token
      - user
    RegisterRequest:
      type: object
      properties:
        nama:
          type: string
          minLength: 1
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 150
        password:
          type: string
          minLength: 1
          writeOnly: true
      required:
      - nama
      - email
      - password
      additionalProperties: true
      description: Field tambahan diabaikan. Registrasi publik selalu role=user dan status=aktif.
      example:
        nama: Pengguna Contoh
        email: user@example.test
        password: contoh-password-lokal
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
          writeOnly: true
      required:
      - email
      - password
      additionalProperties: true
      example:
        email: user@example.test
        password: contoh-password-lokal
    ProfileUpdate:
      type: object
      properties:
        nama:
          type: string
          minLength: 1
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 150
        divisi:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
        jabatan:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
        foto_profil:
          anyOf:
          - type: string
            maxLength: 500
            description: URL http/https; kosong/null menghapus foto. Upload multipart foto menggantikan field ini.
          - type: 'null'
        no_hp:
          type: string
          maxLength: 30
          description: Angka, awalan +, spasi, (), dan -. String kosong menghapus nomor; field yang tidak dikirim
            dipertahankan.
      additionalProperties: false
      minProperties: 1
    ProfileUpload:
      type: object
      properties:
        nama:
          type: string
          minLength: 1
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 150
        divisi:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
        jabatan:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
        foto_profil:
          anyOf:
          - type: string
            maxLength: 500
            description: URL http/https; kosong/null menghapus foto. Upload multipart foto menggantikan field ini.
          - type: 'null'
        no_hp:
          type: string
          maxLength: 30
          description: Angka, awalan +, spasi, (), dan -. String kosong menghapus nomor; field yang tidak dikirim
            dipertahankan.
        foto:
          type: string
          format: binary
          description: Satu file JPG/PNG/WebP, maksimal 2 MiB dan 16 megapiksel. Disimpan ulang sebagai PNG maksimal
            1024×1024.
      additionalProperties: false
      minProperties: 1
...